Switch back to pm.puppetlabs.com for PE downloads
The pm.puppetlabs.com site is available both inside and outside of the firewall
and now includes smart proxying that will detect when pre-releases are being
installed and route accordingly. This is especially important for PE agents
starting with PE 2015.2.0.

Only downside is that pe-releases.puppetlabs.net is faster for users
inside the PL firewall.